+++ lib/vagrant/action/vm/share_folders.rb
@@ -52,7 +52,17 @@ module Vagrant
 
           @env["vm"].ssh.execute do |ssh|
             # short guestpaths first, so we don't step on ourselves
-            shared_folders.sort_by {|name, data| data[:guestpath].length}.each do |name, data|
+            folders = shared_folders.sort_by do |name, data|
+              if data[:guestpath]
+                data[:guestpath].length
+              else
+                # A long enough path to just do this at the end.
+                10000
+              end
+            end
+
+            # Go through each folder and mount
+            folders.each do |name, data|
               if data[:guestpath]
                 # Guest path specified, so mount the folder to specified point
                 @env.ui.info(I18n.t("vagrant.actions.vm.share_folders.mounting_entry",