+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 0.8.8 (unreleased)
 
-
+  - Mount shared folders shortest to longest to avoid mounting
+    subfolders first. [GH-525]
 
 ## 0.8.7 (September 13, 2011)